requirements.txt
`grpcio==1.69.0
grpcio-tools==1.69.0
protobuf==5.29.2
fastapi==0.115.6
uvicorn[standard]==0.34.0
